Actual time resolution
I was looking at the time on the trigger PMT not the time resolution. My code still doesn't quite work, but I have the right idea now. Look into the trigger timing (this might actually be the PMT time) and how to get the PMT timing in a better way.  Current error is "IndexError: too many indices for array"