Complete rewrite
My previous code, which was based off of newbar.py was plotting the wrong data. This code, which still needs to be cleaned up a bit, finds the actual time resolution of the pulses as well as the standard deviation of these values. Most of this is courtesy of timing.py.